Extract code related to Parser into another class
